refactor(git-providers): one directory per provider, one door in
The layout mixed two axes for no designed reason. Transport was
provider-first (`github/http.ts`), capabilities were capability-first
(`content/github.ts`), so `github/client.ts` and `content/github.ts` talked to
the same GitHub, for the same reason, from different places — an artefact of
the order the three PRs were written, not a decision.

Provider is now the primary axis, which turns "the only place GitHub is
hardcoded" from a convention into something you can grep:

  git-providers/
    index.ts          the only entry point the rest of the API imports
    types.ts  content.ts  change-requests.ts    the three contracts
    credentials.ts    which credential reaches which repository
    capabilities.ts   what this deployment can connect
    clients.ts        the composition root — the one module aware of both
    github/  gitlab/  one provider's vocabulary, and nothing else's

Four things were living in the neutral layer that had no business there:

- `env.ts` held both providers' config; it is `github/env.ts` and
  `gitlab/env.ts` now, and each provider answers `capability()` for itself so
  the capabilities tool names no env var and no host.
- The process-wide GitHub App signer sat in the credential ladder. It is in
  `github/app-auth.ts` beside the config it reads, so the ladder — which is
  provider-neutral — holds no GitHub state.
- `repo-choices.ts` rebuilt `https://github.com/owner/repo` by hand and read
  `mcp-github` connections. The legacy half moved to
  `github/legacy-connection.ts`, which hands back a `RepoRef`; the merge is now
  provider-neutral and a GitLab project flows through the same code.
- `findRepositoryForLegacyBinding` assumed `github.com` for its identity
  lookup — correct, but a GitHub fact, so it moved next to the other legacy
  logic. That module is now the single thing that knows a connection can stand
  for a repository, which makes it the whole deletion when the last org
  migrates.

`GIT_ACCOUNT_CONNECT_TOKEN` called `gitlabCurrentUser` directly and hardcoded
"GitHub connects through the App" as a tool-level guard. That is a statement
about what each provider offers, so it is `principalForToken` in the registry
now — and the refusal explains itself: a user PAT would silently widen every
repository to that user's blanket access, where an App mints one scoped to one
repository.

Two reach-throughs remain, both provider-specific BY CONSTRUCTION and both
named in the barrel: `api/routes/git-providers.ts` (an App installation and an
OAuth grant are different redirect dances) and `tools/github/list-user-orgs.ts`
(listing App installations has no counterpart to abstract over).

Pure movement — no behaviour changed. The two factories were merged into one
file and share a `legacyGithubToken` helper, which is what removes the last
copy of the legacy-token path; their different answers to "no credential"
(content throws, change requests return null) are unchanged and now documented
together.
