+++ apps/api/src/git-providers/github/app-auth.ts
@@ -16,7 +16,8 @@ import {
   isPermissionRejected,
   OPTIONAL_MINT_PERMISSIONS,
 } from "@decocms/shared/github-repo-scope";
-import type { GithubAppConfig } from "../env";
+import { type GithubAppConfig, readGithubAppConfig } from "./env";
+import type { GitProviderCapability } from "../types";
 import { GitProviderError } from "../types";
 import {
   githubErrorMessage,
@@ -367,4 +368,41 @@ export class GithubAppAuth {
       if (entries.length < perPage) return installations;
     }
   }
+}
+
+let appAuthSingleton: GithubAppAuth | null | undefined;
+
+/**
+ * The process-wide App signer, or null when this deployment has no App
+ * registered. One instance because its installation-token cache lives inside
+ * it: a second signer would mint a second token per installation and halve
+ * the cache's usefulness.
+ *
+ * Lives here rather than in the credential ladder so that the ladder — which
+ * is provider-neutral — holds no GitHub state of its own.
+ */
+export function getGithubAppAuth(): GithubAppAuth | null {
+  if (appAuthSingleton === undefined) {
+    const config = readGithubAppConfig();
+    appAuthSingleton = config ? new GithubAppAuth(config) : null;
+  }
+  return appAuthSingleton;
+}
+
+/**
+ * The App serves github.com only — GitHub Enterprise needs its own App
+ * registered per instance, which this deployment has no config for. So the
+ * host list is a constant here rather than a setting.
+ */
+const GITHUB_APP_HOST = "github.com";
+
+export function githubCapability(): GitProviderCapability {
+  /**
+   * BOTH a readable config and a signer built from it: a malformed PEM parses
+   * as config and then fails at the first mint, which is a worse answer than
+   * "not configured".
+   */
+  const configured =
+    readGithubAppConfig() !== null && getGithubAppAuth() !== null;
+  return { configured, hosts: configured ? [GITHUB_APP_HOST] : [] };
 }
