+++ apps/api/src/git-providers/gitlab/content.test.ts
@@ -12,8 +12,8 @@ import {
   mapMergeRequestState,
   pooledMap,
   type ResolvedChange,
-} from "./gitlab";
-import { gitlabErrorMessage } from "../gitlab/http";
+} from "./content";
+import { gitlabErrorMessage } from "./http";
 
 describe("buildCommitActions", () => {
   test("an existing file is an update guarded by its last commit", () => {
